# lib/api-zod/src/generated/types/legalMovesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Chess Board API
 * OpenAPI spec version: 0.1.0
 */

export interface LegalMovesResponse {
  square: string;
  moves: string[];
}